fix(workflow): surface malformed activity outputs in eval evidence
activity_result_from_job parsed a runtime job's output JSON with
.ok(), making a malformed payload indistinguishable from no output:
downstream evidence builders reported the step as missing, hiding
producer bugs from operators. Log the deserialization error (with job
id and activity tag) at warn level while keeping the Option contract.

Also split evidence.rs tests into evidence_tests.rs following the
scoring/scoring_tests pattern to stay under the file-size limit, and
add a regression test proving malformed output surfaces as missing
quality-gate evidence rather than a false success.

Closes #1988
